# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Centre Commercial API
  description: |
    API de gestion de centre commercial (MEAN stack).

    ## Rôles
    - **ADMIN** : gestion complète (boutiques, catégories, utilisateurs, stats)
    - **SHOP_MANAGER** : gestion de sa propre boutique (produits, promotions)
    - **CUSTOMER** : navigation, favoris, avis

    ## Authentification
    Stratégie double token JWT :
    - **Access token** : 15 min — à envoyer dans `Authorization: Bearer <token>`
    - **Refresh token** : 7 jours — à envoyer dans le body pour rafraîchir

    Les tokens ne sont délivrés qu'après vérification de l'email.
  version: '1.0.0'
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement
security:
  - BearerAuth: []
...
paths:
...
  /myshop:
    get:
      tags: [My Shop]
      operationId: getMyShop
      summary: Obtenir sa boutique
      responses:
        '200':
          description: Boutique du manager
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [My Shop]
      operationId: updateMyShop
      summary: Modifier sa boutique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShopInput'
      responses:
        '200':
          description: Boutique mise à jour
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Shop'
        '404':
          $ref: '#/components/responses/NotFound'
  /myshop/stats:
    get:
      tags: [My Shop]
      operationId: getMyShopStats
      summary: Statistiques de sa boutique
      responses:
        '200':
          description: Statistiques
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ShopStats'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
...
    UserRole:
      type: string
      enum: [CUSTOMER, SHOP_MANAGER, ADMIN]
    UserStatus:
      type: string
      enum: [VALID, SUSPENDED, WAITING]
    ShopStatus:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED]

    # --- Auth ---
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        address:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    # --- Shops ---
    UpdateShopInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          description: ID de la catégorie
        hours:
          type: string
          example: 'Lun-Sam 9h-19h'
        image:
          type: string
    Shop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        manager:
          $ref: '#/components/schemas/UserProfile'
        category:
          $ref: '#/components/schemas/Category'
        note:
          type: number
          minimum: 0
          maximum: 5
          description: Note moyenne calculée à partir des avis
        hours:
          type: string
        image:
          type: string
        status:
          $ref: '#/components/schemas/ShopStatus'
        createdAt:
          type: string
          format: date-time
    ShopStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalReviews:
          type: integer
        avgRating:
          type: number
...